+++ aws-core/src/main/java/com/amplifyframework/auth/CognitoCredentialsProvider.kt
@@ -16,6 +16,7 @@
 package com.amplifyframework.auth
 
 import aws.smithy.kotlin.runtime.auth.awscredentials.Credentials
+import aws.smithy.kotlin.runtime.util.Attributes
 import com.amplifyframework.AmplifyException
 import com.amplifyframework.core.Amplify
 import com.amplifyframework.core.Consumer
@@ -31,7 +32,7 @@ open class CognitoCredentialsProvider : AuthCredentialsProvider {
     /**
      * Request [Credentials] from the provider.
      */
-    override suspend fun getCredentials(): Credentials {
+    override suspend fun resolve(attributes: Attributes): Credentials {
         return suspendCoroutine { continuation ->
             Amplify.Auth.fetchAuthSession(
                 { authSession ->
@@ -79,7 +80,7 @@ open class CognitoCredentialsProvider : AuthCredentialsProvider {
     override fun getAccessToken(onResult: Consumer<String>, onFailure: Consumer<Exception>) {
         Amplify.Auth.fetchAuthSession(
             { session ->
-                val tokens = session.toAWSAuthSession()?.userPoolTokensResult?.value?.accessToken
+                val tokens = session.toAWSAuthSession()?.accessToken
                 tokens?.let { onResult.accept(tokens) }
                     ?: onFailure.accept(
                         AuthException(
@@ -95,6 +96,6 @@ open class CognitoCredentialsProvider : AuthCredentialsProvider {
     }
 }
 
-private fun AuthSession.toAWSAuthSession(): AWSAuthSessionInternal? {
-    return this as? AWSAuthSessionInternal
+private fun AuthSession.toAWSAuthSession(): AWSAuthSessionBehavior<*>? {
+    return this as? AWSAuthSessionBehavior<*>
 }