+++ maplibre-adapter/src/main/java/com/amplifyframework/geo/maplibre/http/AWSRequestSignerInterceptor.kt
@@ -26,6 +26,7 @@ import aws.smithy.kotlin.runtime.net.QueryParameters
 import aws.smithy.kotlin.runtime.net.Scheme
 import aws.smithy.kotlin.runtime.net.Url
 import aws.smithy.kotlin.runtime.net.toUrlString
+import aws.smithy.kotlin.runtime.util.emptyAttributes
 import com.amplifyframework.geo.location.AWSLocationGeoPlugin
 import java.io.ByteArrayOutputStream
 import java.io.IOException
@@ -54,7 +55,9 @@ internal class AWSRequestSignerInterceptor(
 
         val awsRequest = try {
             // OkHttpRequest -> Signed AWSRequest
-            signRequest(request)
+            runBlocking {
+                signRequest(request)
+            }
         } catch (e: SignCredentialsException) {
             return Response.Builder()
                 .code(401)
@@ -91,7 +94,7 @@ internal class AWSRequestSignerInterceptor(
     }
 
     @Throws(SignCredentialsException::class)
-    private fun signRequest(request: Request): HttpRequest {
+    private suspend fun signRequest(request: Request): HttpRequest {
         val url = request.url
         val headers: AwsHeaders = AwsHeaders.invoke {
             request.headers.forEach { (name, value) ->
@@ -104,7 +107,7 @@ internal class AWSRequestSignerInterceptor(
         val signingConfig = AwsSigningConfig.invoke {
             region = client.config.region
             service = "geo"
-            credentialsProvider = plugin.credentialsProvider
+            credentials = plugin.credentialsProvider.resolve(emptyAttributes())
         }
 
         val httpUrl = Url(